% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efis}
\alias{efis}
\title{EFIS School Board Financial Reports 2021-2024}
\format{
A tibble with 104 rows and 25 columns
}
\usage{
efis
}
\description{
School board financials for Ontario school boards
}
\keyword{datasets}
