% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-clustering.R
\name{fetch_clustering}
\alias{fetch_clustering}
\title{Fetch the current common clustering of t3010 and fir codes}
\usage{
fetch_clustering()
}
\description{
Fetch the current common clustering of t3010 and fir codes
}
