% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-fir.R
\name{fetch_fir}
\alias{fetch_fir}
\title{Fetch FIR data for a given year}
\usage{
fetch_fir(year, dest = NULL)
}
\arguments{
\item{year}{The year for which to fetch FIR data}
}
\value{
A data frame with FIR data for the given year
}
\description{
Fetch FIR data for a given year
}
