% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhin}
\alias{lhin}
\title{LHIN Funding Summary 2017-2021}
\format{
A tibble with 60 rows and 9 columns
}
\usage{
lhin
}
\description{
Local Health Integration Network funding
}
\keyword{datasets}
