% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{on_cpi}
\alias{on_cpi}
\title{Ontario CPI 2015-2024}
\format{
A tibble with 9 rows and 2 columns
}
\usage{
on_cpi
}
\description{
CPI data for All Items in Ontario, 2015-2023, per Statistics Canada
}
\keyword{datasets}
