% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t3010}
\alias{t3010}
\title{Ontario Charity Revenue Data 2019-2023 (T3010)}
\format{
A tibble with 145,645 rows and 23 columns
}
\usage{
t3010
}
\description{
Ontario Charity Revenue Data 2019-2023 (T3010)
}
\keyword{datasets}
